+++ src/models/game_available.rs
@@ -25,7 +25,7 @@ pub struct GameAvailableDTO {
     #[schema(value_type = String, format = DateTime)]
     pub updated_datetime: NaiveDateTime,
     pub status: GameStatus,
-    pub rating: i16,
+    pub rating: u32,
     pub notes: String,
     #[schema(value_type = String, format = Date)]
     pub available_date: NaiveDate,
